refactor: PresetCreateRequest 패키지 이동, final 추가
